cleared out html files and updates README
