+++ manifest.json
@@ -1,7 +1,7 @@
 {
     "name": "4541 Scouting App",
     "short_name": "Scouting App",
-    "start_url": "/compselect.html",
+    "start_url": "/home.html",
     "display": "standalone",
     "background_color": "#fdfdfd",
     "theme_color": "#db4938",
