+++ names.txt
@@ -1,68 +1,11 @@
-Jonah Abellano
-Anthony Avilla
-Taryn Baker
-Logan Bechtel
-Laura Bell
-Brett Boomer
-James Bost
-Lucas Bucht
-Sadie Ceryes
-Brennen Collins
-Dulcey Comeau
-Brady Comeau
-Peter D'Amico
-Paul Davis
-Elijah Davis
-Jay Donyinah
-Gavin Flynn
-Patrick Flynn
-Andrew Gardiner
-Nick Gash
-Jake Goodyear
-Chad Grant
-Kate Greenhill
-Caden Greensfelder
-Ashley Grice
-Thomas Grice
-Bryan Ha
-Grant Hayden
-Reid Haynes
-Max Haynes
-Charles Hecker
-Alex Henn
-Eric Herberich
-Sierra Jablinske
-Matthew Kearney
-Kit Kellam
-Brody Kelly
-Trystan Lam
-Evan Larson-=
-Kayley Lawrence
-Rosie Levin
-Aubrie Lisiewski
-Max Maginnis
-Maggie Malta
-Sam Malta
-Marques Mateo
-Sean McCabe
-Aiden McManus
-Dylan Meacham
-Addisyn Moore
-Savannah Moore
-Aliyah Ali Moulta
-Kate Obara
-Andrew Phelps
-Jacob Phelps
-Kaleb Richardson
-Aeryn Robinson
-Nolan Russ
-Dakota Ryan
-Nicholas Sayan
-Robert Snell
-Grace Stifflemire
-Matthew Studgeon
-Thomas Gardiner
-Daniel Tracy
-Michael Tsampos
-Michael Webb
-Gabe Yoshida
+Marshall D. Teach
+Jesus Burgess
+Shiryu
+Van Augur
+Avalo Pizarro
+Laffitte
+Catarina Devon
+Sanjuan Wolf
+Vasco Shot
+Doc Q
+Kuzan
