--- static/2023pit/css/pit.css
@@ -1,57 +1,0 @@
-h1 {
-    font-weight: bolder;
-    font-size: 48px;
-    margin-left: 600px;
-    margin-bottom: 50px;
-}
-
-.textboxes {
-    margin-left: 75px;
-    padding-top: 10px;
-    padding-left: 50px;
-    padding-right: 50px;
-    padding-bottom: 80px;
-    width: 1250px;
-    border: 5px solid #ff0000;
-    border-radius: 20px;
-}
-
-.textboxes p {
-    font-weight: bold;
-}
-
-.textarea {
-    background-color: transparent;
-    color: #ffffff;
-    padding: 1em;
-    border-radius: 10px;
-    border: 3px solid #ffffff;
-    outline: solid;
-    font-size: 16px;
-    width: 1200px;
-    height: 165px;
-    font-family: Geneva, Segoe UI, Verdana, sans-serif;
-    resize: none;
-}
-
-.submit {
-    height: 50px;
-    width: 200px;
-    background-color: #ff0000; /* Red */
-    border: none;
-    color: white;
-    float: right;
-    text-align: center;
-    text-decoration: none;
-    font-size: 24px;
-    border-radius: 5px;
-    box-shadow: 0 5px #777;
-}
-
-.submit:hover {background-color: #cf0000}
-
-.submit:active {
-  background-color: #cf0000;
-  box-shadow: 0 0px #666;
-  transform: translateY(5px);
-}
