--- static/css/index.css
@@ -1,38 +1,0 @@
-.button_group {
-    margin-top: 50%;
-    width: 100%;
-    text-align: center;
-}
-
-.button_group button{
-    height: 100px;
-    width: 125px;
-    color: white;
-    font-weight: bold;
-    background-color: #ff0000;
-    border: none;
-    box-shadow: 0 10px #777;
-}
-
-.button_group button:hover{background-color: #cf0000}
-
-.button_group button:active {
-    background-color: #cf0000;
-    box-shadow: 0 0px #666;
-    transform: translateY(10px);
-}
-
-.help {
-    border-top-left-radius: 25px;
-    border-bottom-left-radius: 25px;
-}
-
-.competition {
-    
-}
-
-.leaderboard {
-    border-top-right-radius: 25px;
-    border-bottom-right-radius: 25px;
-}
-
