+++ static/currentyear/css/home.css
@@ -18,14 +18,6 @@
     box-shadow: 0 5px #777;
 }
 
-.submit:hover {background-color: #cf0000}
-
-.submit:active {
-    background-color: #cf0000;
-    box-shadow: 0 0px #666;
-    transform: translateY(5px);
-}
-
 .input {
     background-color: #2a2a2a;
     margin-top: 10px;
