--- templates/2023pit/pit.html
@@ -1,29 +1,0 @@
-{% extends "base.html" %}
-{% block styles %}
-<link rel="stylesheet" href="/static/2023pit/css/pit.css">
-{% endblock %}        
-{% block title %}4541 Scouting App | 2023{% endblock %}
-{% block contents %}
-<div>
-    <h1>Pit Scouting</h1>
-    <br>
-    <div class="textboxes">
-        <label for="textarea">
-            <p>What is their score strategy?</p>
-            <textarea id="textarea1" class="textarea" placeholder="Enter text here..."></textarea>
-        <br>
-        <br>
-        <label for="textarea">
-            <p>What is their defense strategy?</p>
-            <textarea id="textarea2" class="textarea" placeholder="Enter text here..."></textarea>
-        <br>
-        <br>
-        <label for="textarea">
-            <p>Other questions about the robot (Freestyle)</p>
-            <textarea id="textarea3" class="textarea" placeholder="Enter text here..."></textarea>
-        <br>
-        <br>
-        <button class="submit" type="submit">Submit</button>
-    </div>
-</div>
-{% endblock %}
