--- templates/compselect.html
@@ -1,20 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}
-4541 Scouting App | Select Competition
-{% endblock %}
-
-{% block contents %}
-<div>
-<h1 style="margin-left: 5%; margin-top: 10%">Select Scouting App Competition</h1>
-    <div style="font-size: 2em; margin-left: 5%;">
-            {% if competitions %}
-            {% for name, url in competitions.items() %}
-            <p>
-                <a class="light-theme" href="{{ url }}">{{ name }}</a>
-            </p>
-            {% endfor %}
-            {% endif %}
-    </div>
-</div>
-{% endblock %}
