--- templates/help.html
@@ -1,3 +1,0 @@
-{% extends "base.html" %}
-{% block title %}4541 Scouting App{% endblock %}
-{% block contents %}This is the help page.{% endblock %}
