+++ templates/home.html
@@ -1,11 +1,12 @@
 {% extends "base.html" %}
-{% block title %}4541 Scouting App | 2023{% endblock %}
+{% block title %}4541 Scouting App | Current Year{% endblock %}
 {% block styles %}
-<link rel="stylesheet" href="/static/2023/css/home.css">
+<link rel="stylesheet" href="/static/currentyear/css/home.css">
+<link rel="stylesheet" href="/static/css/commons.css">
 {% endblock %}
 {% block scripts %}
-<script src="/static/2023/js/home.js"></script>
-<script src="/static/2023/js/navstamp.js"></script>
+<script src="/static/currentyear/js/home.js"></script>
+<script src="/static/currentyear/js/navstamp.js"></script>
 {% endblock %}
 {% block contents %}
 <div class="content">
