--- templates/index.html
@@ -1,18 +1,0 @@
-{% extends "base.html" %}
-<!-- For making pwa: https://www.freecodecamp.org/news/build-a-pwa-from-scratch-with-html-css-and-javascript/ -->
-{% block styles %}
-<link rel="stylesheet" href="/static/css/index.css">
-{% endblock %}
-{% block title %}4541 Scouting App{% endblock %}
-
-{% block contents%}
-<div class="button_group">
-    <div>
-        <h1>Robotics 4541 Scouting App</h1>
-        <br>
-        <button class="help" onclick="window.location = 'help.html';">Help</button>
-        <button class="competition" onclick="window.location = 'compselect.html';">Competition</button>
-        <button class="leaderboard" onclick="window.location = '';">Leaderboard</button>
-    </div>
-</div>
-{% endblock %}
