--- templates/qrscanner.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% block title %}4541 Scouting App{% endblock %}
-{% block styles %}
-{% endblock %}
-{% block scripts %}
-<script src="/static/2023/js/qrcode.js"></script>
-<script src="/static/2023/js/qrscanner.js"></script>
-<script src="/static/2023/js/navstamp.js"></script>
-{% endblock %}
-{% block contents %}
-<div id="qr">            
-</div>
-<div>
-    <input id="qrtext" />
-    <button id="qrbutton" onclick="gencode()" class="button_red">Generate QR</button>
-</div>
-{% endblock %}
